some re-organization of classes
