+++ zio-http/src/main/scala/zhttp/service/client/experimental/ZConnectionManager.scala
@@ -5,11 +5,18 @@ import io.netty.channel.{Channel, ChannelFuture, ChannelFutureListener}
 import io.netty.handler.codec.http.FullHttpRequest
 import zhttp.http.URL
 import zhttp.service.client.ClientSSLHandler.ClientSSLOptions
+import zhttp.service.client.experimental.handler.{ZClientChannelInitializer, ZClientInboundHandler}
 import zio.{Promise, Ref, Task, ZIO}
 
 import java.net.InetSocketAddress
 import scala.collection.mutable
 
+/*
+  Can hold atomic reference to ZConnectionState comprising of
+    - Timeouts like (idleTimeout,requestTimeout,connectionTimeout etc.)
+    - states like (currentTotal)
+    - Data structures like (idleQueue, waitingRequestQueue etc)
+ */
 case class ZConnectionManager(
   connRef: Ref[mutable.Map[InetSocketAddress, Channel]],
   currentExecRef: mutable.Map[Channel, (Promise[Throwable, Resp], FullHttpRequest)] =
@@ -17,37 +24,42 @@ case class ZConnectionManager(
   boo: Bootstrap,
   zExec: zhttp.service.HttpRuntime[Any],
 ) {
+
   def fetchConnection(jReq: FullHttpRequest): Task[Channel] = {
     for {
       mp  <- connRef.get
-//      _   <- ZIO.effect(println(s"CONNECTION MAP : $mp"))
-      url <- ZIO.fromEither(URL.fromString(jReq.uri()))
-      host = url.host
-      port = url.port.getOrElse(80) match {
-        case -1   => 80
-        case port => port
-      }
-      inetSockAddress <- host match {
-        case Some(h) => Task.succeed(new InetSocketAddress(h, port))
-        case _       => Task.fail(new Exception("error getting host"))
-      }
-      conn <- mp.get(inetSockAddress) match {
+      uriAuthority <- getUriAuthority(jReq)
+      // if already key exists for existing connections re-use it
+      // else build a new connection (channel)
+      conn <- mp.get(uriAuthority) match {
         case Some(c) =>
-          println(s"REUSING CONNECTION for $inetSockAddress")
+          println(s"REUSING CONNECTION for $uriAuthority")
+          // To be tested to check if the channel is currently busy
           if (c.isWritable)
             Task.succeed(c)
           else
-            buildChannel("http", inetSockAddress)
-
+            buildChannel("http", uriAuthority)
         case _ =>
-          buildChannel("http", inetSockAddress)
+          buildChannel("http", uriAuthority)
       }
       _ <- connRef.update { m =>
-        m += (inetSockAddress -> conn)
+        m += (uriAuthority -> conn)
       }
     } yield conn
-
   }
+
+  private def getUriAuthority(jReq: FullHttpRequest) = for {
+    url <- ZIO.fromEither(URL.fromString(jReq.uri()))
+    host = url.host
+    port = url.port.getOrElse(80) match {
+      case -1   => 80
+      case port => port
+    }
+    inetSockAddress <- host match {
+      case Some(h) => Task.succeed(new InetSocketAddress(h, port))
+      case _       => Task.fail(new Exception("error getting host"))
+    }
+  } yield (inetSockAddress)
 
   def buildChannel[R](scheme: String, inetSocketAddress: InetSocketAddress): Task[Channel] = {
     for {
@@ -58,31 +70,35 @@ case class ZConnectionManager(
           ClientSSLOptions.DefaultSSL,
         ),
       )
-//      _    <- ZIO.effect(println(s"for ${jReq.uri()} CONNECTING to ${inetSocketAddress}"))
       (h, p) = (inetSocketAddress.toString.split("/")(0), inetSocketAddress.toString.split(":")(1))
 //      _ <- ZIO.effect(println(s"for ${jReq.uri()} CONNECTING to ${(h, p)}"))
       chf = boo.handler(init).connect(h, p.toInt)
-      _ <- ZIO
-        .effect(
-          chf.addListener(new ChannelFutureListener() {
-            override def operationComplete(future: ChannelFuture): Unit = {
-//              val channel = future.channel()
-//              println(s"CONNECTED USING CHH ID: ${channel.id()}")
-              if (!future.isSuccess()) {
-                println(s"error: ${future.cause().getMessage}")
-                future.cause().printStackTrace()
-              } else {
-//                println("FUTURE SUCCESS");
-              }
-            }
-          }): Unit,
-        )
+      // optional can be removed if not really utilised.
+      _ <- attachHandler(chf)
     } yield chf.channel()
+  }
+
+  def attachHandler(chf: ChannelFuture) = {
+    ZIO
+      .effect(
+        chf.addListener(new ChannelFutureListener() {
+          override def operationComplete(future: ChannelFuture): Unit = {
+            if (!future.isSuccess()) {
+              println(s"error: ${future.cause().getMessage}")
+              future.cause().printStackTrace()
+            } else {
+              //                println("FUTURE SUCCESS");
+            }
+          }
+        }): Unit,
+      )
   }
 
   def getActiveConnections: Task[Int] = for {
     mp  <- connRef.get
   } yield (mp.size)
+
+  // release request ???
 }
 
 object ZConnectionManager {}
