+++ zio-http/src/main/scala/zhttp/service/client/experimental/handler/ZClientInboundHandler.scala
@@ -1,25 +1,26 @@
-package zhttp.service.client.experimental
+package zhttp.service.client.experimental.handler
 
 import io.netty.channel.ChannelHandler.Sharable
 import io.netty.channel.{ChannelHandlerContext, SimpleChannelInboundHandler}
 import io.netty.handler.codec.http.FullHttpRequest
 import zhttp.service.HttpRuntime
-
-//import scala.collection.mutable
+import zhttp.service.client.experimental.{Resp, ZConnectionManager}
 
 /**
  * Handles HTTP response
  */
 @Sharable
 final case class ZClientInboundHandler[R](
-  zExec: HttpRuntime[R],
-  connectionManager: ZConnectionManager,
-) extends SimpleChannelInboundHandler[Resp](false) {
+                                           zExec: HttpRuntime[R],
+                                           connectionManager: ZConnectionManager,
+                                         ) extends SimpleChannelInboundHandler[Resp](false) {
 
   override def channelRead0(ctx: ChannelHandlerContext, clientResponse: Resp): Unit = {
-//    println(s"SimpleChannelInboundHandler SimpleChannelInboundHandler CHANNEL READ CONTEXT ID: ${ctx.channel().id()}")
+    //    println(s"SimpleChannelInboundHandler SimpleChannelInboundHandler CHANNEL READ CONTEXT ID: ${ctx.channel().id()}")
     val prom = connectionManager.currentExecRef(ctx.channel())
-    zExec.unsafeRun(ctx) { prom._1.succeed(clientResponse) }
+    zExec.unsafeRun(ctx) {
+      prom._1.succeed(clientResponse)
+    }
   }
 
   override def exceptionCaught(ctx: ChannelHandlerContext, error: Throwable): Unit = {
@@ -29,7 +30,7 @@ final case class ZClientInboundHandler[R](
   }
 
   override def channelActive(ctx: ChannelHandlerContext): Unit = {
-//    println(s"CHANNEL ACTIVE CONTEXT ID: ${ctx.channel().id()}")
+    //    println(s"CHANNEL ACTIVE CONTEXT ID: ${ctx.channel().id()}")
     val prom = connectionManager.currentExecRef(ctx.channel())
     val jReq = prom._2
     ctx.writeAndFlush(jReq): Unit
